Move execution counter to plan.py [Run CI]
